+++ src/main/java/umc/meme/shop/domain/artist/controller/ArtistController.java
@@ -3,9 +3,7 @@ package umc.meme.shop.domain.artist.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import umc.meme.shop.domain.portfolio.dto.request.CreatePortfolioDto;
 import umc.meme.shop.domain.artist.dto.request.ArtistProfileDto;
-import umc.meme.shop.domain.portfolio.dto.request.UpdatePortfolioDto;
 import umc.meme.shop.domain.review.dto.request.UpdateReviewDto;
 import umc.meme.shop.domain.artist.service.ArtistService;
 import umc.meme.shop.global.SuccessStatus;
@@ -21,12 +19,6 @@ public class ArtistController {
     @PatchMapping("/mypage/{userId}/profile/artist")
     public ApiResponse updateProfile(@PathVariable Long userId, @RequestBody ArtistProfileDto profileDto){
         return ApiResponse.SuccessResponse(SuccessStatus.PROFILE_UPDATE);
-    }
-
-    @Operation(summary = "예약 조회", description = "예약 정보를 조회하는 API입니다.")
-    @GetMapping("/mypage/{userId}/reservation")
-    public ApiResponse getReservation(@PathVariable Long userId){
-        return ApiResponse.SuccessResponse(SuccessStatus.RESERVATION_GET, artistService.getReservation(userId));
     }
 
     @Operation(summary = "리뷰 관리", description = "block 상태를 통해 리뷰 공개 유무를 결정할 수 있는 API입니다.")
