+++ src/main/java/umc/meme/shop/domain/reservation/service/ReservationService.java
@@ -3,23 +3,29 @@ package umc.meme.shop.domain.reservation.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import umc.meme.shop.domain.artist.entity.Artist;
+import umc.meme.shop.domain.artist.repository.ArtistRepository;
 import umc.meme.shop.domain.model.entity.Model;
 import umc.meme.shop.domain.model.repository.ModelRepository;
 import umc.meme.shop.domain.portfolio.entity.Portfolio;
 import umc.meme.shop.domain.portfolio.repository.PortfolioRepository;
 import umc.meme.shop.domain.reservation.dto.request.AlterReservationDto;
 import umc.meme.shop.domain.reservation.dto.request.ReservationRequestDto;
+import umc.meme.shop.domain.reservation.dto.response.ReservationResponseDto;
 import umc.meme.shop.domain.reservation.entity.Reservation;
 import umc.meme.shop.domain.reservation.entity.enums.Status;
 import umc.meme.shop.domain.reservation.repository.ReservationRepository;
 import umc.meme.shop.global.ErrorStatus;
 import umc.meme.shop.global.exception.GlobalException;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class ReservationService {
+    private final ArtistRepository artistRepository;
     private final ModelRepository modelRepository;
     private final PortfolioRepository portfolioRepository;
     private final ReservationRepository reservationRepository;
@@ -54,5 +60,16 @@ public class ReservationService {
                 .orElseThrow(() -> new GlobalException(ErrorStatus.NOT_EXIST_RESERVATION));
         Status status = reservationDto.getStatus();
         reservation.updateReservation(status);
+    }
+
+    //아티스트 예약 조회
+    public List<ReservationResponseDto> getReservation(Long artistId){
+        Artist artist = artistRepository.findById(artistId)
+                .orElseThrow(() -> new GlobalException(ErrorStatus.NOT_EXIST_USER));
+
+        List<Reservation> reservationList = reservationRepository.findByArtist(artist);
+        return reservationList.stream()
+                .map(ReservationResponseDto::from)
+                .collect(Collectors.toList());
     }
 }
